+++ src/org/elixir_lang/ecto/Query.kt
@@ -486,7 +486,7 @@ object Query : ModuleWalker(
     /**
      * Whether the `call` is a `join(...)` macro call
      */
-    private fun isJoin(call: Call, state: ResolveState): Boolean = Join.hasNameArity(call) && resolvesTo(call, state)
+    private fun isJoin(call: Call, state: ResolveState): Boolean = Join.hasNameArity(call) && matches(call, state)
 
     private val logger by lazy { com.intellij.openapi.diagnostic.Logger.getInstance(Query::class.java) }
 }